User registration

+++ base/views.py
@@ -5,6 +5,7 @@ from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout 
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -18,9 +19,12 @@ from .forms import RoomForm
 #]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authentiacated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try: 
@@ -37,14 +41,29 @@ def loginPage(request):
             messages.error(request, 'username OR password does not exist')
 
     
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+def registerPage(request):
+    form = UserCreationForm()
 
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+    
+
+    return render(request, 'base/login_register.htmml',{'form' : form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -86,7 +105,7 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
 
-    if request.user != room.user:
+    if request.user != room.host:
         return HttpResponse('You are not allowed here :( )')
 
     if request.method == 'POST':
@@ -102,6 +121,11 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here :( )')
+
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
